moved memory/devirtualizer.hpp -> utilities/devirtualizer.hpp
