+++ src/hotspot/share/gc/serial/defNewGeneration.inline.hpp
@@ -31,8 +31,8 @@
 #include "gc/shared/genCollectedHeap.hpp"
 #include "gc/shared/genOopClosures.inline.hpp"
 #include "gc/shared/space.inline.hpp"
-#include "memory/devirtualizer.inline.hpp"
 #include "oops/access.inline.hpp"
+#include "utilities/devirtualizer.inline.hpp"
 
 // Methods of protected closure types
