+++ src/hotspot/share/gc/shenandoah/shenandoahMark.inline.hpp
@@ -35,11 +35,11 @@
 #include "gc/shenandoah/shenandoahStringDedup.inline.hpp"
 #include "gc/shenandoah/shenandoahTaskqueue.inline.hpp"
 #include "gc/shenandoah/shenandoahUtils.hpp"
-#include "memory/devirtualizer.inline.hpp"
 #include "memory/iterator.inline.hpp"
 #include "oops/compressedOops.inline.hpp"
 #include "oops/oop.inline.hpp"
 #include "runtime/prefetch.inline.hpp"
+#include "utilities/devirtualizer.inline.hpp"
 #include "utilities/powerOfTwo.hpp"
 
 template <StringDedupMode STRING_DEDUP>
