+++ src/hotspot/share/oops/instanceMirrorKlass.inline.hpp
@@ -28,11 +28,11 @@
 #include "oops/instanceMirrorKlass.hpp"
 
 #include "classfile/javaClasses.hpp"
-#include "memory/devirtualizer.inline.hpp"
 #include "oops/instanceKlass.inline.hpp"
 #include "oops/klass.hpp"
 #include "oops/oop.inline.hpp"
 #include "utilities/debug.hpp"
+#include "utilities/devirtualizer.inline.hpp"
 #include "utilities/globalDefinitions.hpp"
 #include "utilities/macros.hpp"
