+++ src/hotspot/share/runtime/stackChunkFrameStream.inline.hpp
@@ -29,13 +29,13 @@
 #include "compiler/oopMap.hpp"
 #include "interpreter/interpreter.hpp"
 #include "logging/log.hpp"
-#include "memory/devirtualizer.inline.hpp"
 #include "oops/method.hpp"
 #include "oops/oop.hpp"
 #include "oops/stackChunkOop.inline.hpp"
 #include "oops/instanceStackChunkKlass.inline.hpp"
 #include "runtime/frame.inline.hpp"
 #include "utilities/debug.hpp"
+#include "utilities/devirtualizer.inline.hpp"
 #include "utilities/globalDefinitions.hpp"
 #include "utilities/macros.hpp"
 #include CPU_HEADER_INLINE(stackChunkFrameStream)
