+++ src/hotspot/share/utilities/devirtualizer.hpp
